// packages/frontend/src/generated/types/InventoryAdjustmentResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AdjustmentType } from './AdjustmentType.ts'

/**
 * @description Response DTO for the `inventory_adjustment` entity.
 */
export interface InventoryAdjustmentResponse {
  /**
   * @type string
   */
  adjustmentType: AdjustmentType
  /**
   * @type string
   */
  amount: string
  /**
   * @type string, uuid
   */
  contractorId: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  productId: string
  /**
   * @type string,null
   */
  reason?: string | null
  /**
   * @type string, uuid
   */
  reconciliationId: string
  /**
   * @type string, uuid
   */
  storageId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}